Summary for a MQGAM fit

Takes a fitted mqgam object produced by qgam::mqgam and produces various useful summaries from it, making use of the mgcv::summary.gam method.

## S3 method for class 'mqgam'
summary(mqgam)

Arguments

mqgam
An mqgam object created with qgam::mqgam.

Value

Please see summary.gam for a detailed explanation of the values given.

Details

Please see summary.gam for a detailed explanation of the summaries given.

References

Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast calibrated additive quantile regression. URL: https://arxiv.org/abs/1707.03307

Wood, S.N. (2011). Fast stable restricted maximum likelihood and marginal likelihood estimation of semiparametric generalized linear models. Journal of the Royal Statistical Society (B) 73(1), 3-36.

Author

D. Schmitz

Examples

# general usage
summary(mqgam = mtqgam_mqgam)
#> $`========================= quantile 0.1 =========================`
#> 
#> Family: elf 
#> Link function: identity 
#> 
#> Formula:
#> dependent ~ s(numeric_1) + s(numeric_2) + factor_1 + factor_2 * 
#>     factor_3
#> 
#> Parametric coefficients:
#>                     Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)           11.852      2.550   4.648 3.36e-06 ***
#> factor_1K              0.781      1.274   0.613   0.5400    
#> factor_1L              8.450     10.603   0.797   0.4255    
#> factor_1M              1.337      1.360   0.983   0.3256    
#> factor_2b             -5.787      3.029  -1.910   0.0561 .  
#> factor_2c             -1.783      3.256  -0.548   0.5840    
#> factor_2d             -2.436      3.023  -0.806   0.4204    
#> factor_2e             -5.077      2.891  -1.756   0.0791 .  
#> factor_3B             -2.908      2.419  -1.202   0.2293    
#> factor_2b:factor_3B    5.294      3.265   1.621   0.1049    
#> factor_2c:factor_3B    1.094      3.474   0.315   0.7529    
#> factor_2d:factor_3B    3.149      3.274   0.962   0.3362    
#> factor_2e:factor_3B    7.011      3.144   2.230   0.0257 *  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Approximate significance of smooth terms:
#>                edf Ref.df Chi.sq p-value  
#> s(numeric_1) 1.021  1.041  1.662  0.2092  
#> s(numeric_2) 3.445  4.269 13.186  0.0115 *
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> R-sq.(adj) =  0.000297   Deviance explained = 69.2%
#> -REML =  24691  Scale est. = 1         n = 5000
#> 
#> $`========================= quantile 0.3 =========================`
#> 
#> Family: elf 
#> Link function: identity 
#> 
#> Formula:
#> dependent ~ s(numeric_1) + s(numeric_2) + factor_1 + factor_2 * 
#>     factor_3
#> 
#> Parametric coefficients:
#>                     Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)         30.90829    3.28923   9.397   <2e-16 ***
#> factor_1K            0.09366    1.73020   0.054   0.9568    
#> factor_1L            9.16071    7.38443   1.241   0.2148    
#> factor_1M            1.23581    1.82688   0.676   0.4987    
#> factor_2b           -5.89874    4.06093  -1.453   0.1463    
#> factor_2c           -4.01288    4.19548  -0.956   0.3388    
#> factor_2d           -3.45826    3.90769  -0.885   0.3762    
#> factor_2e           -6.78947    4.15173  -1.635   0.1020    
#> factor_3B           -3.30077    3.12025  -1.058   0.2901    
#> factor_2b:factor_3B  5.17050    4.36099   1.186   0.2358    
#> factor_2c:factor_3B  3.76976    4.48158   0.841   0.4003    
#> factor_2d:factor_3B  4.07557    4.21973   0.966   0.3341    
#> factor_2e:factor_3B  8.53886    4.44078   1.923   0.0545 .  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Approximate significance of smooth terms:
#>                edf Ref.df Chi.sq p-value  
#> s(numeric_1) 1.008  1.015  0.814  0.3716  
#> s(numeric_2) 2.120  2.644 10.147  0.0164 *
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> R-sq.(adj) =  0.000398   Deviance explained = 30.6%
#> -REML =  24236  Scale est. = 1         n = 5000
#> 
#> $`========================= quantile 0.5 =========================`
#> 
#> Family: elf 
#> Link function: identity 
#> 
#> Formula:
#> dependent ~ s(numeric_1) + s(numeric_2) + factor_1 + factor_2 * 
#>     factor_3
#> 
#> Parametric coefficients:
#>                     Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)          51.9919     2.9191  17.811   <2e-16 ***
#> factor_1K            -0.1199     1.5381  -0.078    0.938    
#> factor_1L             3.5204     6.8065   0.517    0.605    
#> factor_1M             1.0064     1.6229   0.620    0.535    
#> factor_2b            -4.2839     3.6189  -1.184    0.237    
#> factor_2c            -3.3268     3.7782  -0.881    0.379    
#> factor_2d            -2.9186     3.4921  -0.836    0.403    
#> factor_2e            -4.2748     3.6650  -1.166    0.243    
#> factor_3B            -2.7146     2.7831  -0.975    0.329    
#> factor_2b:factor_3B   3.9441     3.8857   1.015    0.310    
#> factor_2c:factor_3B   3.1910     4.0314   0.792    0.429    
#> factor_2d:factor_3B   3.2132     3.7707   0.852    0.394    
#> factor_2e:factor_3B   5.4905     3.9271   1.398    0.162    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Approximate significance of smooth terms:
#>                edf Ref.df Chi.sq p-value   
#> s(numeric_1) 1.008  1.016  0.572 0.45323   
#> s(numeric_2) 2.096  2.615 12.105 0.00682 **
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> R-sq.(adj) =  0.000799   Deviance explained = 0.376%
#> -REML =  23886  Scale est. = 1         n = 5000
#> 
#> $`========================= quantile 0.7 =========================`
#> 
#> Family: elf 
#> Link function: identity 
#> 
#> Formula:
#> dependent ~ s(numeric_1) + s(numeric_2) + factor_1 + factor_2 * 
#>     factor_3
#> 
#> Parametric coefficients:
#>                     Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)          73.1379     3.2990  22.170   <2e-16 ***
#> factor_1K            -0.3210     1.7582  -0.183    0.855    
#> factor_1L            -0.2357     6.8615  -0.034    0.973    
#> factor_1M             0.9733     1.8528   0.525    0.599    
#> factor_2b            -3.3366     4.1706  -0.800    0.424    
#> factor_2c            -3.1010     4.3187  -0.718    0.473    
#> factor_2d            -2.5485     3.9674  -0.642    0.521    
#> factor_2e            -2.0474     4.1850  -0.489    0.625    
#> factor_3B            -2.3422     3.1481  -0.744    0.457    
#> factor_2b:factor_3B   3.3776     4.4735   0.755    0.450    
#> factor_2c:factor_3B   3.1537     4.6090   0.684    0.494    
#> factor_2d:factor_3B   2.6359     4.2848   0.615    0.538    
#> factor_2e:factor_3B   2.8761     4.4822   0.642    0.521    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Approximate significance of smooth terms:
#>                edf Ref.df Chi.sq p-value   
#> s(numeric_1) 1.022  1.044  0.148 0.71879   
#> s(numeric_2) 1.821  2.274 10.319 0.00936 **
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> R-sq.(adj) =  0.000425   Deviance explained = 30.2%
#> -REML =  24268  Scale est. = 1         n = 5000
#> 
#> $`========================= quantile 0.9 =========================`
#> 
#> Family: elf 
#> Link function: identity 
#> 
#> Formula:
#> dependent ~ s(numeric_1) + s(numeric_2) + factor_1 + factor_2 * 
#>     factor_3
#> 
#> Parametric coefficients:
#>                     Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)          91.4252     2.3589  38.757  < 2e-16 ***
#> factor_1K            -0.7171     1.3579  -0.528  0.59745    
#> factor_1L           -12.0658     4.3284  -2.788  0.00531 ** 
#> factor_1M             0.2290     1.4176   0.162  0.87169    
#> factor_2b             1.3651     3.0343   0.450  0.65278    
#> factor_2c            -1.3422     3.0806  -0.436  0.66306    
#> factor_2d            -1.5505     2.7862  -0.556  0.57788    
#> factor_2e            -2.1636     2.8654  -0.755  0.45019    
#> factor_3B            -1.0444     2.2424  -0.466  0.64137    
#> factor_2b:factor_3B  -1.7951     3.2600  -0.551  0.58187    
#> factor_2c:factor_3B   1.0285     3.3011   0.312  0.75539    
#> factor_2d:factor_3B   0.1367     3.0412   0.045  0.96416    
#> factor_2e:factor_3B   1.4764     3.1070   0.475  0.63465    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Approximate significance of smooth terms:
#>                edf Ref.df Chi.sq p-value   
#> s(numeric_1) 1.015  1.031  4.148 0.04288 * 
#> s(numeric_2) 1.056  1.111  9.575 0.00233 **
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> R-sq.(adj) =  -0.00251   Deviance explained = 68.9%
#> -REML =  24687  Scale est. = 1         n = 5000
#> 

# printing just one quantile summary
summary(mqgam = mtqgam_mqgam)[[3]]
#> 
#> Family: elf 
#> Link function: identity 
#> 
#> Formula:
#> dependent ~ s(numeric_1) + s(numeric_2) + factor_1 + factor_2 * 
#>     factor_3
#> 
#> Parametric coefficients:
#>                     Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)          51.9919     2.9191  17.811   <2e-16 ***
#> factor_1K            -0.1199     1.5381  -0.078    0.938    
#> factor_1L             3.5204     6.8065   0.517    0.605    
#> factor_1M             1.0064     1.6229   0.620    0.535    
#> factor_2b            -4.2839     3.6189  -1.184    0.237    
#> factor_2c            -3.3268     3.7782  -0.881    0.379    
#> factor_2d            -2.9186     3.4921  -0.836    0.403    
#> factor_2e            -4.2748     3.6650  -1.166    0.243    
#> factor_3B            -2.7146     2.7831  -0.975    0.329    
#> factor_2b:factor_3B   3.9441     3.8857   1.015    0.310    
#> factor_2c:factor_3B   3.1910     4.0314   0.792    0.429    
#> factor_2d:factor_3B   3.2132     3.7707   0.852    0.394    
#> factor_2e:factor_3B   5.4905     3.9271   1.398    0.162    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Approximate significance of smooth terms:
#>                edf Ref.df Chi.sq p-value   
#> s(numeric_1) 1.008  1.016  0.572 0.45323   
#> s(numeric_2) 2.096  2.615 12.105 0.00682 **
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> R-sq.(adj) =  0.000799   Deviance explained = 0.376%
#> -REML =  23886  Scale est. = 1         n = 5000